chore(useresizeobserver): convert tests to RTL
